only clone on prod build
